Clean up Reminder Backend
