--- ReminderBackEnd/ReminderBackEnd/Data/Contract/Login/DoesUserExistsDatabaseContract.swift
@@ -1,12 +1,0 @@
-//
-//  DoesUserExistsDatabaseContract.swift
-//  Reminder
-//
-//  Created by Arun Kumar on 07/03/22.
-//
-
-import Foundation
-
-public protocol DoesUserExistsDatabaseContract {
-    func doesUserExists(username: String, success: (String) -> Void, failure: (String) -> Void)
-}
