--- ReminderBackEnd/ReminderBackEnd/Data/Contract/Login/GetAllUsersDatabaseContract.swift
@@ -1,12 +1,0 @@
-//
-//  GetAllUsersDatabaseContract.swift
-//  Reminder
-//
-//  Created by Arun Kumar on 08/03/22.
-//
-
-import Foundation
-
-public protocol GetAllUsersDatabaseContract {
-    func getAllUsers(success: ([User]) -> Void, failure: (String) -> Void)
-}
