+++ ReminderBackEnd/ReminderBackEnd/Data/Contract/ReminderDatabaseContract.swift
@@ -1,0 +1,8 @@
+//
+//  ReminderDatabaseContract.swift
+//  ReminderBackEnd
+//
+//  Created by Arun Kumar on 16/03/22.
+//
+
+import Foundation
