+++ ReminderBackEnd/ReminderBackEnd/Domain/Contract/Login/ChangePasswordDataManagerContract.swift
@@ -8,5 +8,5 @@
 import Foundation
 
 public protocol ChangePasswordDataManagerContract {
-    func changePassword(username: String, password: String, success: (String) -> Void, failure: (ChangePasswordError) -> Void)
+    func changePassword(username: String, password: String, success: @escaping (String) -> Void, failure: @escaping (ChangePasswordError) -> Void)
 }
