+++ ReminderWorkspace.xcworkspace/xcuserdata/arun-pt4306.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -126,8 +126,8 @@
             filePath = "Reminder/Reminder/Presentation/Presenter/AppLoginPresenter.swift"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "68"
-            endingLineNumber = "68"
+            startingLineNumber = "67"
+            endingLineNumber = "67"
             landmarkName = "getLastLoggedInUser(onSuccess:onFailure:)"
             landmarkType = "7">
          </BreakpointContent>
@@ -142,8 +142,8 @@
             filePath = "Reminder/Reminder/Presentation/Presenter/AppLoginPresenter.swift"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "66"
-            endingLineNumber = "66"
+            startingLineNumber = "65"
+            endingLineNumber = "65"
             landmarkName = "getLastLoggedInUser(onSuccess:onFailure:)"
             landmarkType = "7">
          </BreakpointContent>
@@ -700,8 +700,8 @@
             filePath = "Reminder/Reminder/Presentation/Presenter/AppLoginPresenter.swift"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "48"
-            endingLineNumber = "48"
+            startingLineNumber = "47"
+            endingLineNumber = "47"
             landmarkName = "getAllUsers(onSuccess:onFailure:)"
             landmarkType = "7">
          </BreakpointContent>
